+++ Management/Models/Local/Assignment/RubricItem.cs
@@ -4,6 +4,6 @@ public record RubricItem
 {
   public static readonly string extraCredit = "(Extra Credit) ";
   public required string Label { get; set; }
-  public required int Points { get; set; }
+  public required double Points { get; set; }
   public bool IsExtraCredit => Label.Contains(extraCredit.ToLower(), StringComparison.CurrentCultureIgnoreCase);
 }
